added logging

+++ src/data/download_data.py
@@ -1,15 +1,21 @@
 import csv
 import requests
+import logging
+from pathlib import Path
 from src import config
 
-url = config.price_paid_url    
-
-fname = config.project_home + "\\src\\data\\price_paid.csv"
+url = config.price_paid_url
+    
+fname = Path(config.project_home).joinpath("data","external","price_paid.csv")
 
 def download_data(url = url, file_to_write = fname):
-    print("Making get request from {}".format(url))
+    if fname.exists():
+        logging.warn("File already exists at {filename}. Data won't be downloaded".format(filename=str(fname)))
+        return None
+    logging.info("Making get request from {}".format(url))
+
     response = requests.get(url, stream=True)
-    print("Writing file to {}".format(file_to_write))
+    logging.info("Writing file to {}".format(file_to_write))
     print_after = 1000000
     i=0
     with open(fname, 'w') as f:
@@ -18,4 +24,4 @@ def download_data(url = url, file_to_write = fname):
             writer.writerow(line.decode('utf-8').split(','))
             i+=1
             if i%print_after==0:
-                print("Written {} lines...".format(i))
+                logging.info("Written {} lines...".format(i))
